+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 **Other changes:**
 
 - Removed reference to the ``.A`` attribute and replaced it with ``.toarray()``.
+- Add support between formulaic and pandas 3.0
 
 4.0.0 - 2024-04-23
 ------------------