+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Unreleased
 
 - Removed reference to the ``.A`` attribute and replaced it with ``.toarray()``.
 - Add support between formulaic and pandas 3.0
+- Support pypi release for numpy 2.0
 
 4.0.0 - 2024-04-23
 ------------------